refactor: Simplify library structure, consolidate classes, and add QIIME2 integration tests
Consolidates all modular classes (`SeqSample`, `SeqSampleNotPaired`,
`SeqSamplePairedNotInterleaved`, `ItsPosition`, `Dedup`) into a single file
`itsxpress/SeqSample.py`. Removes redundant modular files and updates
imports/__init__.py. Moves the QIIME2 integration tests to a dedicated optional
test module `tests/test_qiime2_integration.py` which automatically skips when
`qiime2` is not installed. Fixes coordinate-zero, generator-exhaustion,
and IndexError edge cases. Adds extensive type hinting, Google-style docstrings,
and new targeted standalone unit tests. Keep the three analytical markdown
documents at the repo root.
